Update processed data
This commit introduces changes to the data processing pipeline
to more easily enable the ETL of sample data into the SQLite
database. Concretely, the code that formats the final ContentItem
lists for each platform is moved into normalize_posts.py and is
used by preprocessing.py to write the newline-delimited JSON
files for each platform. In turn, both the data puller and db
seeder scripts use these JSON files to, respectively, generate
a sample user feed, or to load the data in the SQLite.

The API for data puller CLI was not changed.